+++ include/hicma/operations/misc/get_dim.h
@@ -6,17 +6,19 @@
 #include "yorel/yomm2/cute.hpp"
 using yorel::yomm2::virtual_;
 
+#include <cstdint>
+
 
 namespace hicma
 {
 
-int get_n_rows(const Node&);
+int64_t get_n_rows(const Node&);
 
-declare_method(int, get_n_rows_omm, (virtual_<const Node&>))
+declare_method(int64_t, get_n_rows_omm, (virtual_<const Node&>))
 
-int get_n_cols(const Node&);
+int64_t get_n_cols(const Node&);
 
-declare_method(int, get_n_cols_omm, (virtual_<const Node&>))
+declare_method(int64_t, get_n_cols_omm, (virtual_<const Node&>))
 
 } // namespace hicma
