+++ src/operations/arithmetic/subtraction.cpp
@@ -7,6 +7,7 @@
 #include "hicma/operations/misc/get_dim.h"
 
 #include <cassert>
+#include <cstdint>
 
 
 namespace hicma
@@ -20,8 +21,8 @@ NodeProxy operator-(const Node& A, const Node& B) {
 
 define_method(NodeProxy, subtraction_omm, (const Dense& A, const Dense& B)) {
   Dense out(A.dim[0], A.dim[1]);
-  for (int i=0; i<A.dim[0]; i++) {
-    for (int j=0; j<A.dim[1]; j++) {
+  for (int64_t i=0; i<A.dim[0]; i++) {
+    for (int64_t j=0; j<A.dim[1]; j++) {
       out(i, j) = A(i, j) - B(i, j);
     }
   }
